pyproject.toml
[project]
name = "self-improving-coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=1.0.7",
    "litellm>=1.44.0",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.0",
    "psycopg[binary]>=3.1.0",
    "pgvector>=0.2.5",
    "mem0ai>=1.0.2",
    "neo4j>=5.12.0",
    "opentelemetry-api>=1.26.0",
    "opentelemetry-sdk>=1.26.0",
    "opentelemetry-exporter-otlp>=1.26.0",
    "prometheus-client>=0.20.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "ruff>=0.14.14",
    "bandit>=1.9.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "agent",
    "app",
    "tools",
    "rag",
    "memory",
    "observability",
    "llm",
    "eval",
    "storage",
    "runs",
]

